TIP-860: move phpunit tests
