+++ tests/back/Pim/Enrichment/Integration/Doctrine/Query/CountImpactedProductsIntegration.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Pim\Bundle\EnrichBundle\tests\integration\Doctrine\ORM\Query;
+namespace AkeneoTest\Pim\Enrichment\Integration\Doctrine\Query;
 
 use Akeneo\Test\Integration\Configuration;
 use Akeneo\Test\Integration\TestCase;
