+++ tests/back/Pim/Enrichment/Integration/ProductQueryBuilder/Sorter/InGroupSorterIntegration.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\EnrichBundle\tests\integration\PQB\Sorter;
+namespace AkeneoTest\Pim\Enrichment\Integration\ProductQueryBuilder\Sorter;
 
 use Akeneo\Pim\Enrichment\Bundle\tests\Integration\PQB\AbstractProductQueryBuilderTestCase;
 use Akeneo\Pim\Enrichment\Component\Product\Query\Sorter\Directions;
